// DXC_Frontend/src/api/models/feedbackProcessingDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface FeedbackProcessingDto {
  id?: number
  publicId?: string
  fromStatusId?: number
  /** @nullable */
  fromStatusCode: string | null
  /** @nullable */
  fromStatusName: string | null
  toStatusId?: number
  /** @nullable */
  toStatusCode: string | null
  /** @nullable */
  toStatusName: string | null
  /** @nullable */
  assignedDepartmentId?: number | null
  /** @nullable */
  assignedDepartmentCode?: string | null
  /** @nullable */
  assignedDepartmentName?: string | null
  /** @nullable */
  assignedByUserPublicId?: string | null
  assignedAt?: string
  /** @nullable */
  processingNote?: string | null
  createdAt?: string
}